memfs: Add thread-safety to Memory Make Memory thread safe for operations such as Create, Rename and Delete. When compared with Memory without a mutex, the cost per op is roughly 50ns, as per benchmark below:
goos: linux
goarch: amd64
pkg: github.com/go-git/go-billy/v6/test
cpu: AMD Ryzen 7 PRO 8840HS w/ Radeon 780M Graphics
BenchmarkCompare/stdlib_open-16         	  341277	      3303 ns/op	     152 B/op	       3 allocs/op
BenchmarkCompare/osfs.chrootOS_open-16  	  286580	      3806 ns/op	     360 B/op	       9 allocs/op
BenchmarkCompare/osfs.boundOS_open-16   	  209929	      5311 ns/op	     984 B/op	      20 allocs/op
BenchmarkCompare/memfs_open-16          	 1503024	       967.4 ns/op	     224 B/op	      10 allocs/op
BenchmarkCompare/memfs_mutex_open-16    	 1272890	       911.6 ns/op	     224 B/op	      10 allocs/op
BenchmarkCompare/stdlib_read-16         	  184956	      5957 ns/op	 171.88 MB/s	       0 B/op	       0 allocs/op
BenchmarkCompare/osfs.chrootOS_read-16  	  180552	      6207 ns/op	 164.97 MB/s	       0 B/op	       0 allocs/op
BenchmarkCompare/osfs.boundOS_read-16   	  179314	      6131 ns/op	 167.02 MB/s	       0 B/op	       0 allocs/op
BenchmarkCompare/memfs_read-16          	  792040	      1568 ns/op	 653.02 MB/s	       0 B/op	       0 allocs/op
BenchmarkCompare/memfs_mutex_read-16    	  783422	      1609 ns/op	 636.27 MB/s	       0 B/op	       0 allocs/op
BenchmarkCompare/stdlib_create-16       	  216717	      5232 ns/op	     152 B/op	       3 allocs/op
BenchmarkCompare/osfs.chrootOS_create-16         	  157268	      7017 ns/op	     616 B/op	      11 allocs/op
BenchmarkCompare/osfs.boundOS_create-16          	  123038	      9242 ns/op	    1335 B/op	      22 allocs/op
BenchmarkCompare/memfs_create-16                 	  880813	      1954 ns/op	     693 B/op	      12 allocs/op
BenchmarkCompare/memfs_mutex_create-16           	  559534	      1901 ns/op	     631 B/op	      12 allocs/op
PASS
ok  	github.com/go-git/go-billy/v6/test	118.368s

This change enables users to opt-out (memfs.WithoutMutex), in case speed
on a single thread is more important than being thread-safe.

Additional thread-safety will be required for memfs.file.
