+++ .github/workflows/test.yml
@@ -2,7 +2,6 @@ name: Test
 
 on: 
   push:
-    branches: [ "master", "main" ]
   pull_request:
 
 permissions: {}
