+++ Makefile
@@ -14,7 +14,7 @@ $(GOLANGCI):
 
 .PHONY: test
 test:
-	$(GOTEST) -race -timeout 60s ./...
+	$(GOTEST) -race -timeout 300s ./...
 
 test-coverage:
 	echo "" > $(COVERAGE_REPORT); \
