+++ memfs/memory.go
@@ -4,7 +4,6 @@ package memfs // import "github.com/go-git/go-billy/v6/memfs"
 import (
 	"errors"
 	"fmt"
-	"io"
 	"io/fs"
 	"log"
 	"os"
@@ -12,7 +11,6 @@ import (
 	"sort"
 	"strings"
 	"syscall"
-	"time"
 
 	"github.com/go-git/go-billy/v6"
 	"github.com/go-git/go-billy/v6/helper/chroot"
@@ -24,11 +22,25 @@ const separator = filepath.Separator
 // Memory a very convenient filesystem based on memory files.
 type Memory struct {
 	s *storage
+
+	m mutex
 }
 
 // New returns a new Memory filesystem.
-func New() billy.Filesystem {
-	fs := &Memory{s: newStorage()}
+func New(opts ...Option) billy.Filesystem {
+	o := &options{
+		// Enable thread-safety by default.
+		newMutex: newMutex,
+	}
+
+	for _, opt := range opts {
+		opt(o)
+	}
+
+	fs := &Memory{
+		s: newStorage(o.newMutex),
+		m: o.newMutex(),
+	}
 	_, err := fs.s.New("/", 0755|os.ModeDir, 0)
 	if err != nil {
 		log.Printf("failed to create root dir: %v", err)
@@ -215,172 +227,6 @@ func (fs *Memory) Capabilities() billy.Capability {
 		billy.ReadAndWriteCapability |
 		billy.SeekCapability |
 		billy.TruncateCapability
-}
-
-type file struct {
-	name     string
-	content  *content
-	position int64
-	flag     int
-	mode     os.FileMode
-	modTime  time.Time
-
-	isClosed bool
-}
-
-func (f *file) Name() string {
-	return f.name
-}
-
-func (f *file) Read(b []byte) (int, error) {
-	n, err := f.ReadAt(b, f.position)
-	f.position += int64(n)
-
-	if errors.Is(err, io.EOF) && n != 0 {
-		err = nil
-	}
-
-	return n, err
-}
-
-func (f *file) ReadAt(b []byte, off int64) (int, error) {
-	if f.isClosed {
-		return 0, os.ErrClosed
-	}
-
-	if !isReadAndWrite(f.flag) && !isReadOnly(f.flag) {
-		return 0, errors.New("read not supported")
-	}
-
-	n, err := f.content.ReadAt(b, off)
-
-	return n, err
-}
-
-func (f *file) Seek(offset int64, whence int) (int64, error) {
-	if f.isClosed {
-		return 0, os.ErrClosed
-	}
-
-	switch whence {
-	case io.SeekCurrent:
-		f.position += offset
-	case io.SeekStart:
-		f.position = offset
-	case io.SeekEnd:
-		f.position = int64(f.content.Len()) + offset
-	}
-
-	return f.position, nil
-}
-
-func (f *file) Write(p []byte) (int, error) {
-	return f.WriteAt(p, f.position)
-}
-
-func (f *file) WriteAt(p []byte, off int64) (int, error) {
-	if f.isClosed {
-		return 0, os.ErrClosed
-	}
-
-	if !isReadAndWrite(f.flag) && !isWriteOnly(f.flag) {
-		return 0, errors.New("write not supported")
-	}
-
-	f.modTime = time.Now()
-	n, err := f.content.WriteAt(p, off)
-	f.position = off + int64(n)
-
-	return n, err
-}
-
-func (f *file) Close() error {
-	if f.isClosed {
-		return os.ErrClosed
-	}
-
-	f.isClosed = true
-	return nil
-}
-
-func (f *file) Truncate(size int64) error {
-	if size < int64(len(f.content.bytes)) {
-		f.content.bytes = f.content.bytes[:size]
-	} else if more := int(size) - len(f.content.bytes); more > 0 {
-		f.content.bytes = append(f.content.bytes, make([]byte, more)...)
-	}
-
-	return nil
-}
-
-func (f *file) Duplicate(filename string, mode fs.FileMode, flag int) billy.File {
-	nf := &file{
-		name:    filename,
-		content: f.content,
-		mode:    mode,
-		flag:    flag,
-		modTime: f.modTime,
-	}
-
-	if isTruncate(flag) {
-		nf.content.Truncate()
-	}
-
-	if isAppend(flag) {
-		nf.position = int64(nf.content.Len())
-	}
-
-	return nf
-}
-
-func (f *file) Stat() (os.FileInfo, error) {
-	return &fileInfo{
-		name:    f.Name(),
-		mode:    f.mode,
-		size:    f.content.Len(),
-		modTime: f.modTime,
-	}, nil
-}
-
-// Lock is a no-op in memfs.
-func (f *file) Lock() error {
-	return nil
-}
-
-// Unlock is a no-op in memfs.
-func (f *file) Unlock() error {
-	return nil
-}
-
-type fileInfo struct {
-	name    string
-	size    int
-	mode    os.FileMode
-	modTime time.Time
-}
-
-func (fi *fileInfo) Name() string {
-	return fi.name
-}
-
-func (fi *fileInfo) Size() int64 {
-	return int64(fi.size)
-}
-
-func (fi *fileInfo) Mode() fs.FileMode {
-	return fi.mode
-}
-
-func (fi *fileInfo) ModTime() time.Time {
-	return fi.modTime
-}
-
-func (fi *fileInfo) IsDir() bool {
-	return fi.mode.IsDir()
-}
-
-func (*fileInfo) Sys() interface{} {
-	return nil
 }
 
 func (c *content) Truncate() {
