+++ memfs/storage.go
@@ -1,40 +1,42 @@
 package memfs
 
 import (
-	"errors"
 	"fmt"
-	"io"
 	"io/fs"
 	"os"
 	"path/filepath"
 	"strings"
-	"sync"
 	"time"
+
+	"github.com/go-git/go-billy/v6"
 )
 
 type storage struct {
 	files    map[string]*file
 	children map[string]map[string]*file
+
+	mf mutex
+	mc mutex
 }
 
-func newStorage() *storage {
+func newStorage(newMutex func() mutex) *storage {
 	return &storage{
 		files:    make(map[string]*file, 0),
 		children: make(map[string]map[string]*file, 0),
+		mc:       newMutex(),
+		mf:       newMutex(),
 	}
 }
 
 func (s *storage) Has(path string) bool {
-	path = clean(path)
-
-	_, ok := s.files[path]
+	_, ok := s.get(path)
 	return ok
 }
 
 func (s *storage) New(path string, mode fs.FileMode, flag int) (*file, error) {
 	path = clean(path)
-	if s.Has(path) {
-		if !s.MustGet(path).mode.IsDir() {
+	if f, ok := s.get(path); ok {
+		if !f.mode.IsDir() {
 			return nil, fmt.Errorf("file already exists %q", path)
 		}
 
@@ -42,7 +44,6 @@ func (s *storage) New(path string, mode fs.FileMode, flag int) (*file, error) {
 	}
 
 	name := filepath.Base(path)
-
 	f := &file{
 		name:    name,
 		content: &content{name: name},
@@ -51,7 +52,10 @@ func (s *storage) New(path string, mode fs.FileMode, flag int) (*file, error) {
 		modTime: time.Now(),
 	}
 
+	s.mf.Lock()
 	s.files[path] = f
+	s.mf.Unlock()
+
 	err := s.createParent(path, mode, f)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create parent: %w", err)
@@ -71,27 +75,32 @@ func (s *storage) createParent(path string, mode fs.FileMode, f *file) error {
 		return err
 	}
 
+	s.mc.Lock()
 	if _, ok := s.children[base]; !ok {
 		s.children[base] = make(map[string]*file, 0)
 	}
 
 	s.children[base][f.Name()] = f
+	s.mc.Unlock()
+
 	return nil
 }
 
 func (s *storage) Children(path string) []*file {
 	path = clean(path)
 
-	l := make([]*file, 0)
+	s.mc.RLock()
+	l := make([]*file, 0, len(s.children))
 	for _, f := range s.children[path] {
 		l = append(l, f)
 	}
+	s.mc.RUnlock()
 
 	return l
 }
 
 func (s *storage) MustGet(path string) *file {
-	f, ok := s.Get(path)
+	f, ok := s.get(path)
 	if !ok {
 		panic(fmt.Errorf("couldn't find %q", path))
 	}
@@ -100,12 +109,19 @@ func (s *storage) MustGet(path string) *file {
 }
 
 func (s *storage) Get(path string) (*file, bool) {
+	return s.get(path)
+}
+
+func (s *storage) get(path string) (*file, bool) {
 	path = clean(path)
-	if !s.Has(path) {
+
+	s.mf.RLock()
+	file, ok := s.files[path]
+	s.mf.RUnlock()
+	if !ok {
 		return nil, false
 	}
 
-	file, ok := s.files[path]
 	return file, ok
 }
 
@@ -113,12 +129,16 @@ func (s *storage) Rename(from, to string) error {
 	from = clean(from)
 	to = clean(to)
 
+	if from == "/" || from == "." {
+		return billy.ErrBaseDirCannotBeRenamed
+	}
+
 	if !s.Has(from) {
 		return os.ErrNotExist
 	}
 
 	move := [][2]string{{from, to}}
-
+	s.mf.RLock()
 	for pathFrom := range s.files {
 		if pathFrom == from || !strings.HasPrefix(pathFrom, from) {
 			continue
@@ -129,6 +149,7 @@ func (s *storage) Rename(from, to string) error {
 
 		move = append(move, [2]string{pathFrom, pathTo})
 	}
+	s.mf.RUnlock()
 
 	for _, ops := range move {
 		from := ops[0]
@@ -143,104 +164,63 @@ func (s *storage) Rename(from, to string) error {
 }
 
 func (s *storage) move(from, to string) error {
+	s.mf.Lock()
 	s.files[to] = s.files[from]
 	s.files[to].name = filepath.Base(to)
+	file := s.files[to]
+	s.mf.Unlock()
+
+	s.mc.Lock()
 	s.children[to] = s.children[from]
+	s.mc.Unlock()
 
 	defer func() {
-		delete(s.children, from)
+		s.mf.Lock()
 		delete(s.files, from)
+		s.mf.Unlock()
+
+		s.mc.Lock()
+		delete(s.children, from)
 		delete(s.children[filepath.Dir(from)], filepath.Base(from))
+		s.mc.Unlock()
 	}()
 
-	return s.createParent(to, 0644, s.files[to])
+	return s.createParent(to, 0644, file)
 }
 
 func (s *storage) Remove(path string) error {
 	path = clean(path)
+	if path == "/" || path == "." {
+		return billy.ErrBaseDirCannotBeRemoved
+	}
 
-	f, has := s.Get(path)
+	f, has := s.get(path)
 	if !has {
 		return os.ErrNotExist
 	}
 
-	if f.mode.IsDir() && len(s.children[path]) != 0 {
-		return fmt.Errorf("dir: %s contains files", path)
+	if f.mode.IsDir() {
+		s.mc.RLock()
+		if len(s.children[path]) != 0 {
+			s.mc.RUnlock()
+			return fmt.Errorf("dir: %s contains files", path)
+		}
+		s.mc.RUnlock()
 	}
 
 	base, file := filepath.Split(path)
 	base = filepath.Clean(base)
 
-	delete(s.children[base], file)
+	s.mf.Lock()
 	delete(s.files, path)
+	s.mf.Unlock()
+
+	s.mc.Lock()
+	delete(s.children[base], file)
+	s.mc.Unlock()
 	return nil
 }
 
 func clean(path string) string {
 	return filepath.Clean(filepath.FromSlash(path))
-}
-
-type content struct {
-	name  string
-	bytes []byte
-
-	m sync.RWMutex
-}
-
-func (c *content) WriteAt(p []byte, off int64) (int, error) {
-	if off < 0 {
-		return 0, &os.PathError{
-			Op:   "writeat",
-			Path: c.name,
-			Err:  errors.New("negative offset"),
-		}
-	}
-
-	c.m.Lock()
-	prev := len(c.bytes)
-
-	diff := int(off) - prev
-	if diff > 0 {
-		c.bytes = append(c.bytes, make([]byte, diff)...)
-	}
-
-	c.bytes = append(c.bytes[:off], p...)
-	if len(c.bytes) < prev {
-		c.bytes = c.bytes[:prev]
-	}
-	c.m.Unlock()
-
-	return len(p), nil
-}
-
-func (c *content) ReadAt(b []byte, off int64) (n int, err error) {
-	if off < 0 {
-		return 0, &os.PathError{
-			Op:   "readat",
-			Path: c.name,
-			Err:  errors.New("negative offset"),
-		}
-	}
-
-	c.m.RLock()
-	size := int64(len(c.bytes))
-	if off >= size {
-		c.m.RUnlock()
-		return 0, io.EOF
-	}
-
-	l := int64(len(b))
-	if off+l > size {
-		l = size - off
-	}
-
-	btr := c.bytes[off : off+l]
-	n = copy(b, btr)
-
-	if len(btr) < len(b) {
-		err = io.EOF
-	}
-	c.m.RUnlock()
-
-	return
 }
